# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/platform/whatsapp-templates/sync": {
      "post": {
        "operationId": "WhatsappTemplatesController_sync",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sincronizar catálogo de templates a partir da WABA da plataforma",
        "tags": [
          "Platform — WhatsApp Templates"
        ]
      }
    },
    "/platform/whatsapp-templates": {
      "get": {
        "operationId": "WhatsappTemplatesController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar templates do catálogo (inclui slots parseados)",
        "tags": [
          "Platform — WhatsApp Templates"
        ]
      }
    },
    "/platform/whatsapp-templates/{id}/test": {
      "post": {
        "operationId": "WhatsappTemplatesController_test",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestWhatsappTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enviar template do catálogo para um número (sem TenantLead/coin)",
        "tags": [
          "Platform — WhatsApp Templates"
        ]
      }
    },
...
    "/tenant/push-subscriptions": {
      "put": {
        "operationId": "PushSubscriptionsController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPushSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Subscription registrada ou atualizada"
          },
          "401": {
            "description": "JWT ausente ou inválido"
          },
          "403": {
            "description": "Tenant inativo"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar ou atualizar PushSubscription do usuário autenticado",
        "tags": [
          "Tenant — Push"
        ]
      },
      "delete": {
        "operationId": "PushSubscriptionsController_remove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletePushSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Subscription removida"
          },
          "401": {
            "description": "JWT ausente ou inválido"
          },
          "403": {
            "description": "Tenant inativo"
          },
          "404": {
            "description": "Subscription não encontrada para este usuário"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover PushSubscription do usuário autenticado",
        "tags": [
          "Tenant — Push"
        ]
      }
    }
  },
  "info": {
    "title": "Gestão de Leads",
    "description": "API para gestão de leads",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TestWhatsappTemplateDto": {
        "type": "object",
        "properties": {
          "to": {
            "type": "string",
            "minLength": 1,
            "description": "Telefone de destino (dígitos; 55 será prefixado se faltar)",
            "example": "11999999999"
          },
          "variables": {
            "type": "object",
            "description": "Valores por slot key (literal ou tipo de binding)",
            "example": {
              "body.1": "Demo"
            },
            "default": {}
          },
          "leadId": {
            "type": "number",
            "minimum": 1,
            "description": "Lead opcional para resolver bindings lead.*",
            "example": 1
          }
        },
        "required": [
          "to"
        ]
      },
...
      "PushSubscriptionKeysDto": {
        "type": "object",
        "properties": {
          "p256dh": {
            "type": "string",
            "description": "Chave p256dh da PushSubscription do browser"
          },
          "auth": {
            "type": "string",
            "description": "Chave auth da PushSubscription do browser"
          }
        },
        "required": [
          "p256dh",
          "auth"
        ]
      },
      "UpsertPushSubscriptionDto": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string",
            "description": "Endpoint único retornado pelo PushManager.subscribe()"
          },
          "keys": {
            "$ref": "#/components/schemas/PushSubscriptionKeysDto"
          }
        },
        "required": [
          "endpoint",
          "keys"
        ]
      },
      "DeletePushSubscriptionDto": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string",
            "description": "Endpoint da subscription a remover"
          }
        },
        "required": [
          "endpoint"
        ]
      }
    }
  }
}